(frontend) Ir files are built in the truth output folder
Fixes #5. Now each time an IR file is serialized from an Actor/Unit,
the file is generated in the output folder corresponding to the project
where the CAL file live. To achieve this, we had to store the project
of AstActor/AstUnit in an attribute of the corresponding Actor/Unit