(tests) Remove myjar from tests
On the CI server, the test suite is run completely outside of
eclipse/osgi platform. There is no concept of bundle to build the
classpath. To test extraction of files/folder to the disk, we use
a special library which is always added to the classpath: org.jgrapht

This should fix issues with unit tests on the server