(C backend) Adds explicit type casting when necessary
If the expression being assigned to a variable with `=` is a binary
expression and it contains sub-expressions of different bit sizes
e.g. one of the two sub-expressions is an unsigned 32bit expression
and the outermost binary expression a 64bit expression, then this adds
explicit type casting to the outer most expression before assigning
its value to the left hand side target variable. For example

    c = (u64) a * b;
    d = (u32) c / 8;

Fixes orcc/orcc#135 .